feat: asset publishing no longer depends on asset hash (#65)
The previous mechanism for publishing assets depended on the stability of the asset hash between the local `cdk synth` and the Synthesize step in github actions. Long story short, this asset hash is actually quite brittle and can break due to differing versions of node. The current workaround is to remove CDK Metadata from the template, since that is the resource that encodes the node version.

This PR aims to remove asset hashes from `deploy.yaml` entirely, so as to remove the dependency to this hash. During `cdk synth`, we now write a bash script per asset to `cdk.out`, and reference this script in the github step. We will now always find the asset we are looking for.

A side-effect is that `deploy.yaml` no longer changes when the asset changes, since we don't depend on the hash.