feat: github openId construct and ability to authenticate to aws via openId  (#81)
This PR adds an openId option for authenticating against AWS. Previously the only way to authenticate against AWS was to provide long-lived github secrets with AWS credentials. Now we can configure a role that uses an oidc identity provider to authenticate.

This PR adds two functionalities:

A CDK construct `GitHubActionRole` that configures the necessary role for you, and returns the role arn.

A new property in `GitHubWorkflow` called `gitHubActionRoleArn` that takes in the role arn and if set, the workflow will authenticate via role rather than github secrets.

The role that we create via `GitHubActionRole` has the necessary permissions to assume CDK bootstrapped roles and login to ecr.