feat: diff protection (#98)
Adds diff protection between local `cdk synth` and the `cdk synth` performed via GitHub Actions. To achieve this, I've added a known environment variable, `SYNTH` that is set to `'github'` during the workflow. If this environment variable is set, then we check to make sure that the `.yml` file that was committed is equal to the `.yml` file that gets generated.

TODO: add an escape hatch for users who want to disable this feature.