fix: synth job should have oidc permissions (#182)
Fixes #174. Continuation of #172.

The Synthesize job does not use the OIDC role out of the box, so there was no need to configure permissions to allow it in that job. However, the job includes `preBuildSteps`, which are user-configured steps that could rely on the OIDC role to work (main use case: live lookups during synth time). It's not really recommended to introduce non-determinism in this way (see discussion on #174) but it needs to be supported nevertheless.

Co-authored-by: @awshilpa