fix: remove session token property when undefined (#201)
Previously we would always pass in the session token to authenticate whether or not it was necessary. When no session token secret was specified, we would pass in `aws-session-token: \${{ secrets.undefined }}`.

I guess it did not break any users, but it's ugly and unnecessary. This PR removes it, and only adds the session token parameter when necessary.