chore: add a 'default' case to the pipeline type switch (#344)
In https://github.com/aws/aws-cdk/pull/21619, we are making the exhaustiveness check for `GraphAnnotation` impossible on purpose, so that all CDK Pipelines implementations will be forced to add a `default` case to the `switch` (in preparation of future extensions of the graph model).

Add a `default` case here already to make sure we don't fail compilation.

Fixes #