feat: specify name and/or url in gitHubEnvironment (#518)
Extended typing for `AddGitHubStageOptions.githubEnvironment` to allow for specifying an environment url. 

No changes to the underlying code were required.

Added a snapshot test to verify that the YAML file is created in accordance with  https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-using-environment-name-and-url

BREAKING CHANGE: `gitHubEnvironment` is now typed as a `GitHubEnvironment` and not `string`.