feat: option to configure partition with env var (#895)
Adds support for deployment to aws-cn partition by checking an optional environemt variable, `CDK_AWS_PARTITION`. The partition is currently not available to the pipeline in the `jobForDeploy`, adding the environment variable is a non-breaking change to get the pipeline to deploy to aws-cn.

This has been verified with stacks and nested stacks in multiple stages in aws-cn.

Fixes #820